backend updated

+++ backend/EventPlanning/src/main/java/spring/event/repository/LoginRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import spring.event.model.Login;
 
-public interface LoginRepository extends CrudRepository<Login,Long> {
+public interface LoginRepository extends CrudRepository<Login,Integer> {
 
 	
 	Login findByUsername(String name);
