+++ backend/EventPlanning/src/main/java/spring/event/repository/RegistrationRepository.java
@@ -1,13 +1,13 @@
 package spring.event.repository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
 
-import spring.event.model.Login;
+import org.springframework.data.repository.CrudRepository;
+
 import spring.event.model.Users;
 
-public interface RegistrationRepository extends CrudRepository<Users,Long> {
-	Users findByEmail(String email);
-    Users findByUserid(long userid);
+public interface RegistrationRepository extends CrudRepository<Users,Integer>{
+	  
+	    Users findByEmail(String email);
+	    Users findByUserid(long userid);
+	     
 
 }
